mobile/android/actors/tests/mochitests/test_geckoview_actor_telemetry.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1816151
-->
<head>
  <meta charset="utf-8">
  <title>Tests for Telemetry in GeckoView Actors</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="head.js" type="application/javascript"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1816151">Mozilla Bug 1816151 for Window.Print() Telemetry</a>
<script class="testbody" type="text/javascript">

  const printScript = SpecialPowers.loadChromeScript(_ => {
    /* eslint-env mozilla/chrome-script */
    addMessageListener("print",() => {
      const navigator = Services.wm.getMostRecentWindow("navigator:geckoview");
      const printActor = navigator.window.moduleManager.getActor("GeckoViewPrintDelegate");
      printActor.telemetryDotPrintRequested();
    });
  });

  add_task(async function test_windowDotPrintTelemetry() {
    const telemetryStart = await GleanTest.dotprint.requested.testGetValue() ?? 0;

    // Using the print actor directly because
    // if window.print() is requested Android exits the TestRunnerActivity and starts a PrintActivity,
    // which causes the test harness to stop unexpectedly
    await printScript.sendAsyncMessage("print");

    const requestPrintOnce = await GleanTest.dotprint.requested.testGetValue() ?? 0;
    is(requestPrintOnce - telemetryStart, 1, "GeckoView Dot Print Telemetry Incremented Once");

    await printScript.sendAsyncMessage("print");
    const requestPrintTwice = await GleanTest.dotprint.requested.testGetValue() ?? 0;
    is(requestPrintTwice - telemetryStart, 2, "GeckoView Dot Print Telemetry Incremented Twice");
  });

</script>
</body>
</html>
